requirements.txt
Flask==2.3.3
flask-cors==4.0.1
gunicorn==21.2.0
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
rapidfuzz==3.10.1
python-dotenv==1.0.1
stripe
